@Document needed on implementing records - not on Card
replace Test page by RegistrationPage.vue
