+++ frontend/src/feature/category/pages/CategoryOverviewPage.vue
@@ -23,63 +23,6 @@ const items = ref([] as
   { category: Category, expanded: boolean }[]
 );
 
-// ------------------- TEST DATA -------------------
-// ---uncomment if on npm run mode - and comment out the fetchCategories() call
-// items.value.push({category: {id: "123", name: "test", description: "test", numberOfCards: 1, actions:
-//   [
-//     {
-//       "action": ActionType.WRITE,
-//       "uri": "http://localhost:8080/categories/123"
-//     },
-//     {
-//       "action": ActionType.DELETE,
-//       "uri": "http://localhost:8080/categories/123"
-//     }
-//   ]
-//   }, expanded: false});
-//
-// items.value.push({category: {id: "1234", name: "test2", description: "test description", numberOfCards: 2, actions:
-//       [
-//         {
-//           "action": ActionType.READ,
-//           "uri": "http://localhost:8080/categories/123"
-//         },
-//         {
-//           "action": ActionType.DELETE,
-//           "uri": "http://localhost:8080/categories/123"
-//         }
-//       ]
-//   }, expanded: false});
-//
-// items.value.push({category: {id: "12345", name: "test3", description: "test with blah", numberOfCards: 3, actions:
-//       [
-//         {
-//           "action": ActionType.WRITE,
-//           "uri": "http://localhost:8080/categories/123"
-//         },
-//         {
-//           "action": ActionType.READ,
-//           "uri": "http://localhost:8080/categories/123"
-//         }
-//       ]
-//   }, expanded: false});
-// items.value.push({category: {id: "123456", name: "test-last", description: "test", numberOfCards: 2, actions:
-//       [
-//         {
-//           "action": ActionType.READ,
-//           "uri": "http://localhost:8080/categories/123"
-//         },
-//         {
-//           "action": ActionType.WRITE,
-//           "uri": "http://localhost:8080/categories/123"
-//         },
-//         {
-//           "action": ActionType.DELETE,
-//           "uri": "http://localhost:8080/categories/123"
-//         }
-//       ]
-//   }, expanded: false});
-
 const fetchCategories = async () => {
   items.value = [];
   const categories = await useCategoriesService().getCategories();
