+++ frontend/src/layouts/SideNavigation.vue
@@ -3,7 +3,7 @@
     <v-list-item title="Flashcards" subtitle="Menu"></v-list-item>
     <v-divider></v-divider>
     <v-list-item to="/categories" title="Your card categories"></v-list-item>
-    <v-list-item to="/test" title="Test axios"></v-list-item>
+    <v-list-item to="/register" title="User registration"></v-list-item>
     <v-divider></v-divider>
     <v-list-item to="/" title="Welcome"></v-list-item>
   </v-navigation-drawer>
