+++ server/src/main/java/org/hyperskill/community/flashcards/card/model/MultipleChoiceQuiz.java
@@ -6,6 +6,7 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.PersistenceCreator;
 import org.springframework.data.annotation.Transient;
 import org.springframework.data.annotation.TypeAlias;
+import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.time.Instant;
 import java.util.List;
@@ -18,6 +19,7 @@ import java.util.Set;
  */
 @Builder
 @TypeAlias("MCQ")
+@Document
 public record MultipleChoiceQuiz(
     @Id String id,
     String title,
@@ -32,6 +34,7 @@ public record MultipleChoiceQuiz(
     public MultipleChoiceQuiz setPermissions(String permissions) {
         return new MultipleChoiceQuiz(id, title, tags, question, createdAt, permissions, options, correctOptions);
     }
+
     @PersistenceCreator
     @SuppressWarnings("unused")
     public MultipleChoiceQuiz(String id, String title, Set<String> tags, String question, List<String> options,
