+++ server/src/main/java/org/hyperskill/community/flashcards/card/model/QuestionAndAnswer.java
@@ -6,12 +6,14 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.PersistenceCreator;
 import org.springframework.data.annotation.Transient;
 import org.springframework.data.annotation.TypeAlias;
+import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.time.Instant;
 import java.util.Set;
 
 @Builder
 @TypeAlias("QNA")
+@Document
 public record QuestionAndAnswer(
         @Id String id,
         String title,
